requirements.txt
selenium == 4.27.1
webdriver-manager == 4.0.2
pillow == 11.1.0 
pytesseract == 0.3.13
